requirements.txt
numpy~=1.22.1
pillow~=9.0.0